Firebase and InvoiceUser Page added

+++ pages/invoice-user.tsx
@@ -1,32 +1,70 @@
 "use client";
 import {useEffect, useState} from "react";
-import {collection, addDoc, getDoc, query, onSnapshot, DocumentData, doc, deleteDoc} from 'firebase/firestore';
-import {db} from "@/firebase/config";
-import {InvoiceUser} from "@/types/general";
-import {BsFillTrashFill} from "react-icons/bs";
+import {
+    collection,
+    DocumentData,
+    doc,
+    deleteDoc,
+    setDoc,
+    updateDoc, query, onSnapshot
+} from 'firebase/firestore';
+import {BsFillTrashFill, BsPencilSquare} from "react-icons/bs";
 import Layout from "../components/layout";
+import {InvoiceUser} from "@/src/types/general";
+import InvoiceUserModal from "@/components/modals/InvoiceUserModal";
+import {db} from "@/src/firebase/config";
+
+export const emptyInvoiceUser = {
+    supplierName: "",
+    supplierTaxNumber: "",
+    supplierPostCode: "",
+    supplierTown: "",
+    supplierStreetName: "",
+    supplierStreet: "",
+    supplierAddress: "",
+    supplierBankAccountNumber: "",
+
+    login: "",
+    password: "",
+    signKey: "",
+    exchangeKey: "",
+};
 
 export default function InvoiceUser() {
-    // TODO: Make Root layout to be working here
     const [users, setUsers] = useState([] as InvoiceUser[]);
-    const [newUser, setNewUser] = useState({} as InvoiceUser);
+    const [currentUser, setCurrentUser] = useState({...emptyInvoiceUser} as InvoiceUser);
+    const [showNewUser, setShowNewUser] = useState(false);
     const dbName = 'invoice_users';
 
-    const addUser = async (e) => {
-        e.preventDefault();
-        if (newUser.login) {
-            await addDoc(collection(db, dbName), newUser as DocumentData);
-            setNewUser({});
+    const saveUser = async () => {
+        if (currentUser.id) {
+            // If there is an ID we need to update
+            const userRef = doc(db, dbName, currentUser.id);
+            await updateDoc(userRef, currentUser as DocumentData);
+            setCurrentUser({...emptyInvoiceUser});
+            setShowNewUser(false);
+        } else if (currentUser.supplierName) {
+            const userRef = doc(collection(db, dbName));
+            await setDoc(userRef, currentUser as DocumentData, {merge: true});
+            setCurrentUser({...emptyInvoiceUser});
+            setShowNewUser(false);
+        } else {
+            alert('Name is mandatory field');
         }
     };
 
     const deleteUser = async (id) => {
-        await deleteDoc(doc(db, dbName, id));
+        if (id && window.confirm('Are you sure you wish to delete this User?')) {
+            await deleteDoc(doc(db, dbName, id));
+        }
+    };
+
+    const openUser = (user: InvoiceUser) => {
+        setCurrentUser(Object.assign({}, emptyInvoiceUser, user));
+        setShowNewUser(true);
     };
 
     useEffect(() => {
-        // TODO: Env Variables -> Auth -> Test this method
-        /*
         const q = query(collection(db, dbName));
         const unsubscribe = onSnapshot(q, (querySnapshot) => {
             const receivedUsers = [];
@@ -35,12 +73,23 @@ export default function InvoiceUser() {
             });
             setUsers(receivedUsers);
             return () => unsubscribe()
-        })*/
+        })
     }, []);
 
     return (
         <Layout>
-            <div className="relative overflow-x-auto shadow-md sm:rounded-lg max-w-screen-xl w-full mt-4">
+            <div className="flex justify-between">
+                <div />
+                <button type="button"
+                        className="text-white bg-gray-800 hover:bg-gray-900 focus:outline-none
+                            focus:ring-4 focus:ring-gray-300 font-medium rounded-lg text-sm px-5 py-2.5 mr-2
+                            dark:bg-gray-800 dark:hover:bg-gray-700 dark:focus:ring-gray-700 dark:border-gray-700"
+                        onClick={() => setShowNewUser(true)}
+                >
+                    Create User
+                </button>
+            </div>
+            <div className="relative overflow-x-auto shadow-md sm:rounded-lg max-w-screen-xl w-full mt-2">
                 <table className="w-full text-sm text-left text-gray-500 dark:text-gray-400">
                     <thead className="text-xs text-gray-700 uppercase bg-gray-50 dark:bg-gray-700 dark:text-gray-400">
                     <tr>
@@ -48,7 +97,10 @@ export default function InvoiceUser() {
                             ID
                         </th>
                         <th scope="col" className="px-6 py-3">
-                            login
+                            Name
+                        </th>
+                        <th scope="col" className="px-6 py-3">
+                            Login
                         </th>
                         <th scope="col" className="px-6 py-3">
                             Action
@@ -65,9 +117,14 @@ export default function InvoiceUser() {
                             </th>
                             <th scope="row"
                                 className="px-6 py-4 font-medium text-gray-900 whitespace-nowrap dark:text-white">
+                                {user.supplierName}
+                            </th>
+                            <th scope="row"
+                                className="px-6 py-4 font-medium text-gray-900 whitespace-nowrap dark:text-white">
                                 {user.login}
                             </th>
                             <td className="px-6 py-4 flex flex-row text-lg">
+                                <BsPencilSquare className="cursor-pointer ml-2" onClick={() => openUser(user)}/>
                                 <BsFillTrashFill className="cursor-pointer ml-2" onClick={() => deleteUser(user.id)}/>
                             </td>
                         </tr>
@@ -75,6 +132,13 @@ export default function InvoiceUser() {
                     </tbody>
                 </table>
             </div>
+            <InvoiceUserModal
+                visible={showNewUser}
+                onClose={() => setShowNewUser(false)}
+                onSave={() => saveUser()}
+                currentUser={currentUser}
+                setCurrentUser={setCurrentUser}
+            />
         </Layout>
 
     );
